+++ src/apify_client/client.py
@@ -1,7 +1,6 @@
 from typing import Dict, Optional, Union
 
 from ._http_client import _HTTPClient, _HTTPClientAsync
-from ._utils import _make_async_docs
 from .clients import (
     ActorClient,
     ActorClientAsync,
@@ -272,7 +271,6 @@ class ApifyClientAsync(_BaseApifyClient):
 
     http_client: _HTTPClientAsync
 
-    @_make_async_docs(src=ApifyClient.__init__)
     def __init__(
         self,
         token: Optional[str] = None,
@@ -282,6 +280,16 @@ class ApifyClientAsync(_BaseApifyClient):
         min_delay_between_retries_millis: Optional[int] = 500,
         timeout_secs: Optional[int] = 360,
     ):
+        """Initialize the ApifyClientAsync.
+
+        Args:
+            token (str, optional): The Apify API token
+            api_url (str, optional): The URL of the Apify API server to which to connect to. Defaults to https://api.apify.com
+            max_retries (int, optional): How many times to retry a failed request at most
+            min_delay_between_retries_millis (int, optional): How long will the client wait between retrying requests
+                (increases exponentially from this value)
+            timeout_secs (int, optional): The socket timeout of the HTTP requests sent to the Apify API
+        """
         super().__init__(
             token,
             api_url=api_url,
@@ -297,90 +305,139 @@ class ApifyClientAsync(_BaseApifyClient):
             timeout_secs=self.timeout_secs,
         )
 
-    @_make_async_docs(src=ApifyClient.actor)
     def actor(self, actor_id: str) -> ActorClientAsync:
+        """Retrieve the sub-client for manipulating a single actor.
+
+        Args:
+            actor_id (str): ID of the actor to be manipulated
+        """
         return ActorClientAsync(resource_id=actor_id, **self._options())
 
-    @_make_async_docs(src=ApifyClient.actors)
     def actors(self) -> ActorCollectionClientAsync:
+        """Retrieve the sub-client for manipulating actors."""
         return ActorCollectionClientAsync(**self._options())
 
-    @_make_async_docs(src=ApifyClient.build)
     def build(self, build_id: str) -> BuildClientAsync:
+        """Retrieve the sub-client for manipulating a single actor build.
+
+        Args:
+            build_id (str): ID of the actor build to be manipulated
+        """
         return BuildClientAsync(resource_id=build_id, **self._options())
 
-    @_make_async_docs(src=ApifyClient.builds)
     def builds(self) -> BuildCollectionClientAsync:
+        """Retrieve the sub-client for querying multiple builds of a user."""
         return BuildCollectionClientAsync(**self._options())
 
-    @_make_async_docs(src=ApifyClient.run)
     def run(self, run_id: str) -> RunClientAsync:
+        """Retrieve the sub-client for manipulating a single actor run.
+
+        Args:
+            run_id (str): ID of the actor run to be manipulated
+        """
         return RunClientAsync(resource_id=run_id, **self._options())
 
-    @_make_async_docs(src=ApifyClient.runs)
     def runs(self) -> RunCollectionClientAsync:
+        """Retrieve the sub-client for querying multiple actor runs of a user."""
         return RunCollectionClientAsync(**self._options())
 
-    @_make_async_docs(src=ApifyClient.dataset)
     def dataset(self, dataset_id: str) -> DatasetClientAsync:
+        """Retrieve the sub-client for manipulating a single dataset.
+
+        Args:
+            dataset_id (str): ID of the dataset to be manipulated
+        """
         return DatasetClientAsync(resource_id=dataset_id, **self._options())
 
-    @_make_async_docs(src=ApifyClient.datasets)
     def datasets(self) -> DatasetCollectionClientAsync:
+        """Retrieve the sub-client for manipulating datasets."""
         return DatasetCollectionClientAsync(**self._options())
 
-    @_make_async_docs(src=ApifyClient.key_value_store)
     def key_value_store(self, key_value_store_id: str) -> KeyValueStoreClientAsync:
+        """Retrieve the sub-client for manipulating a single key-value store.
+
+        Args:
+            key_value_store_id (str): ID of the key-value store to be manipulated
+        """
         return KeyValueStoreClientAsync(resource_id=key_value_store_id, **self._options())
 
-    @_make_async_docs(src=ApifyClient.key_value_stores)
     def key_value_stores(self) -> KeyValueStoreCollectionClientAsync:
+        """Retrieve the sub-client for manipulating key-value stores."""
         return KeyValueStoreCollectionClientAsync(**self._options())
 
-    @_make_async_docs(src=ApifyClient.request_queue)
     def request_queue(self, request_queue_id: str, *, client_key: Optional[str] = None) -> RequestQueueClientAsync:
+        """Retrieve the sub-client for manipulating a single request queue.
+
+        Args:
+            request_queue_id (str): ID of the request queue to be manipulated
+            client_key (str): A unique identifier of the client accessing the request queue
+        """
         return RequestQueueClientAsync(resource_id=request_queue_id, client_key=client_key, **self._options())
 
-    @_make_async_docs(src=ApifyClient.request_queues)
     def request_queues(self) -> RequestQueueCollectionClientAsync:
+        """Retrieve the sub-client for manipulating request queues."""
         return RequestQueueCollectionClientAsync(**self._options())
 
-    @_make_async_docs(src=ApifyClient.webhook)
     def webhook(self, webhook_id: str) -> WebhookClientAsync:
+        """Retrieve the sub-client for manipulating a single webhook.
+
+        Args:
+            webhook_id (str): ID of the webhook to be manipulated
+        """
         return WebhookClientAsync(resource_id=webhook_id, **self._options())
 
-    @_make_async_docs(src=ApifyClient.webhooks)
     def webhooks(self) -> WebhookCollectionClientAsync:
+        """Retrieve the sub-client for querying multiple webhooks of a user."""
         return WebhookCollectionClientAsync(**self._options())
 
-    @_make_async_docs(src=ApifyClient.webhook_dispatch)
     def webhook_dispatch(self, webhook_dispatch_id: str) -> WebhookDispatchClientAsync:
+        """Retrieve the sub-client for accessing a single webhook dispatch.
+
+        Args:
+            webhook_dispatch_id (str): ID of the webhook dispatch to access
+        """
         return WebhookDispatchClientAsync(resource_id=webhook_dispatch_id, **self._options())
 
-    @_make_async_docs(src=ApifyClient.webhook_dispatches)
     def webhook_dispatches(self) -> WebhookDispatchCollectionClientAsync:
+        """Retrieve the sub-client for querying multiple webhook dispatches of a user."""
         return WebhookDispatchCollectionClientAsync(**self._options())
 
-    @_make_async_docs(src=ApifyClient.schedule)
     def schedule(self, schedule_id: str) -> ScheduleClientAsync:
+        """Retrieve the sub-client for manipulating a single schedule.
+
+        Args:
+            schedule_id (str): ID of the schedule to be manipulated
+        """
         return ScheduleClientAsync(resource_id=schedule_id, **self._options())
 
-    @_make_async_docs(src=ApifyClient.schedules)
     def schedules(self) -> ScheduleCollectionClientAsync:
+        """Retrieve the sub-client for manipulating schedules."""
         return ScheduleCollectionClientAsync(**self._options())
 
-    @_make_async_docs(src=ApifyClient.log)
     def log(self, build_or_run_id: str) -> LogClientAsync:
+        """Retrieve the sub-client for retrieving logs.
+
+        Args:
+            build_or_run_id (str): ID of the actor build or run for which to access the log
+        """
         return LogClientAsync(resource_id=build_or_run_id, **self._options())
 
-    @_make_async_docs(src=ApifyClient.task)
     def task(self, task_id: str) -> TaskClientAsync:
+        """Retrieve the sub-client for manipulating a single task.
+
+        Args:
+            task_id (str): ID of the task to be manipulated
+        """
         return TaskClientAsync(resource_id=task_id, **self._options())
 
-    @_make_async_docs(src=ApifyClient.tasks)
     def tasks(self) -> TaskCollectionClientAsync:
+        """Retrieve the sub-client for manipulating tasks."""
         return TaskCollectionClientAsync(**self._options())
 
-    @_make_async_docs(src=ApifyClient.user)
     def user(self, user_id: Optional[str] = None) -> UserClientAsync:
+        """Retrieve the sub-client for querying users.
+
+        Args:
+            user_id (str, optional): ID of user to be queried. If None, queries the user belonging to the token supplied to the client
+        """
         return UserClientAsync(resource_id=user_id, **self._options())