+++ src/apify_client/clients/resource_clients/key_value_store_collection.py
@@ -1,6 +1,6 @@
 from typing import Any, Dict, Optional
 
-from ..._utils import ListPage, _filter_out_none_values_recursively, _make_async_docs
+from ..._utils import ListPage, _filter_out_none_values_recursively
 from ..base import ResourceCollectionClient, ResourceCollectionClientAsync
 
 
@@ -19,7 +19,7 @@ class KeyValueStoreCollectionClient(ResourceCollectionClient):
         limit: Optional[int] = None,
         offset: Optional[int] = None,
         desc: Optional[bool] = None,
-    ) -> ListPage:
+    ) -> ListPage[Dict]:
         """List the available key-value stores.
 
         https://docs.apify.com/api/v2#/reference/key-value-stores/store-collection/get-list-of-key-value-stores
@@ -58,7 +58,6 @@ class KeyValueStoreCollectionClientAsync(ResourceCollectionClientAsync):
         resource_path = kwargs.pop('resource_path', 'key-value-stores')
         super().__init__(*args, resource_path=resource_path, **kwargs)
 
-    @_make_async_docs(src=KeyValueStoreCollectionClient.list)
     async def list(
         self,
         *,
@@ -66,9 +65,32 @@ class KeyValueStoreCollectionClientAsync(ResourceCollectionClientAsync):
         limit: Optional[int] = None,
         offset: Optional[int] = None,
         desc: Optional[bool] = None,
-    ) -> ListPage:
+    ) -> ListPage[Dict]:
+        """List the available key-value stores.
+
+        https://docs.apify.com/api/v2#/reference/key-value-stores/store-collection/get-list-of-key-value-stores
+
+        Args:
+            unnamed (bool, optional): Whether to include unnamed key-value stores in the list
+            limit (int, optional): How many key-value stores to retrieve
+            offset (int, optional): What key-value store to include as first when retrieving the list
+            desc (bool, optional): Whether to sort the key-value stores in descending order based on their modification date
+
+        Returns:
+            ListPage: The list of available key-value stores matching the specified filters.
+        """
         return await self._list(unnamed=unnamed, limit=limit, offset=offset, desc=desc)
 
-    @_make_async_docs(src=KeyValueStoreCollectionClient.get_or_create)
     async def get_or_create(self, *, name: Optional[str] = None, schema: Optional[Dict] = None) -> Dict:
+        """Retrieve a named key-value store, or create a new one when it doesn't exist.
+
+        https://docs.apify.com/api/v2#/reference/key-value-stores/store-collection/create-key-value-store
+
+        Args:
+            name (str, optional): The name of the key-value store to retrieve or create.
+            schema (Dict, optional): The schema of the key-value store
+
+        Returns:
+            dict: The retrieved or newly-created key-value store.
+        """
         return await self._get_or_create(name=name, resource=_filter_out_none_values_recursively({'schema': schema}))