+++ src/apify_client/clients/resource_clients/webhook_dispatch.py
@@ -1,6 +1,5 @@
 from typing import Any, Dict, Optional
 
-from ..._utils import _make_async_docs
 from ..base import ResourceClient, ResourceClientAsync
 
 
@@ -31,6 +30,12 @@ class WebhookDispatchClientAsync(ResourceClientAsync):
         resource_path = kwargs.pop('resource_path', 'webhook-dispatches')
         super().__init__(*args, resource_path=resource_path, **kwargs)
 
-    @_make_async_docs(src=WebhookDispatchClient.get)
     async def get(self) -> Optional[Dict]:
+        """Retrieve the webhook dispatch.
+
+        https://docs.apify.com/api/v2#/reference/webhook-dispatches/webhook-dispatch-object/get-webhook-dispatch
+
+        Returns:
+            dict, optional: The retrieved webhook dispatch, or None if it does not exist
+        """
         return await self._get()