+++ CHANGELOG.md
@@ -43,6 +43,7 @@ Changelog
 - improved type hints across codebase
 - added option to manually publish the package with a workflow dispatch
 - added `pre-commit` to run code quality checks before committing
+- converted `unittest`-style tests to `pytest`-style tests
 
 [0.6.0](../../releases/tag/v0.6.0) - 2022-06-27
 -----------------------------------------------